+++ R/amlr_db.R
@@ -4,28 +4,25 @@
 #' with US AMLR default parameters
 #'
 #' @param Database character; name of database to connect to
+#' @param Server character; name of the database server
 #' @param Driver character; default is 'ODBC Driver 18 for SQL Server'
-#' @param Server character; default is 'swc-***REMOVED***-s'
 #' @param Trusted_Connection character; default is 'Yes'
 #' @param Encrypt character; default is 'Optional'
 #' @param idleTimeout integer; default is one hour
 #' @param silent logical; default is \code{TRUE}.
-#'   Passed as an argument to \code{\link[base]{try}}
-#' @param ... additional arguments passed to \code{\link[pool]{dbPool}} or
-#'   \code{\link[odbc]{dbConnect}}
+#'   Passed as an argument to [base::try()]
+#' @param ... additional arguments passed to [pool::dbPool()] or
+#'   [odbc::dbConnect()]
 #'
 #' @name amlr_db
 #"
-#' @details
-#' Wrapper for a call to \code{\link[pool]{dbPool}} or
-#' \code{\link[odbc]{dbConnect}},
-#' with an \code{\link[odbc]{odbc}} driver passed to \code{drv}.
-#' See \code{\link[pool]{dbPool}} or \code{\link[odbc]{dbConnect}}
-#' for more information about these arguments.
+#' @details Wrapper for a call to [pool::dbPool()] or [odbc::dbConnect()], with
+#' the [odbc::odbc()] driver passed to \code{drv}. See the above-linked
+#' functions for more information about these arguments.
 #'
-#' Arguments that are \code{NULL} are ignored and not passed to
-#' the database connection function. This allows you to, for instance pass a
-#' username and password rather than Trusted_Connection string (see examples).
+#' Arguments that are \code{NULL} are ignored and not passed to the database
+#' connection function. This allows you to, for instance pass a username and
+#' password rather than Trusted_Connection string (see examples).
 #'
 #' @return
 #' Output of \code{\link[base]{try}(\link[pool]{dbPool}(..), silent = TRUE)}, or
@@ -33,22 +30,23 @@
 #'
 #' @examples
 #' \dontrun{
-#' amlr_dbPool("***REMOVED***")
-#' amlr_dbConnect("***REMOVED***")
+#' db.name <- "db-name"
+#' server.name <- "server-name"
+#' amlr_dbPool(Database = db.name, Server = server.name)
+#' amlr_dbConnect(Database = db.name, Server = server.name)
 #'
 #' # Connect using username and password
 #' amlr_dbPool(
-#'   "***REMOVED***",
-#'   Trusted_Connection = NULL, uid = "sa", pwd = "SecurePwd"
+#'   Database, Server,
+#'   Trusted_Connection = NULL, uid = "username", pwd = "SecurePwd"
 #' )
 #' }
 #'
 #' @seealso \url{https://github.com/rstudio/pool}
 #'
 #' @export
-amlr_dbPool <- function(Database,
+amlr_dbPool <- function(Database, Server,
                         Driver = "ODBC Driver 18 for SQL Server",
-                        Server = "swc-***REMOVED***-s",
                         Trusted_Connection = "Yes",
                         Encrypt = "Optional",
                         idleTimeout = 3600000,
@@ -68,9 +66,8 @@ amlr_dbPool <- function(Database,
 
 #' @name amlr_db
 #' @export
-amlr_dbConnect <- function(Database,
+amlr_dbConnect <- function(Database, Server,
                            Driver = "ODBC Driver 18 for SQL Server",
-                           Server = "swc-***REMOVED***-s",
                            Trusted_Connection = "Yes",
                            Encrypt = "Optional",
                            silent = TRUE,
@@ -89,8 +86,7 @@ amlr_dbConnect <- function(Database,
 
 # Check driver
 .driver_check <- function(Driver) {
-  driver.rec <- "ODBC Driver 18 for SQL Server"
-  if (Driver != driver.rec)
-    warning("You are not connecting with the recommended driver: ", driver.rec,
-            immediate. = TRUE)
+  if (Driver != amlrDatabases::amlr.driver)
+    warning("You are not connecting with the AMLR-preferred driver: ",
+            amlrDatabases::amlr.driver, immediate. = TRUE)
 }