+++ R/data.R
@@ -1,0 +1,7 @@
+#' AMLR-preferred database driver
+#'
+#' A variable that contains the name of the preferred driver for connecting to
+#' US AMLR SQL Server databases
+#'
+#' @format A character vector of length one
+"amlr.driver"