+++ R/mod_database.R
@@ -1,39 +1,48 @@
-#' Database module for AMLR shiny apps
+#' Database connection module for AMLR shiny apps
 #'
-#' Database module for AMLR shiny apps
+#' Database connection module for AMLR shiny apps
 #'
 #' @name mod_database
 #'
-#' @param id module namespace, see \code{shiny::\link[shiny]{NS}}
+#' @param id module namespace, see [shiny::NS()]
 #' @param col.width integer; column width of column of UI widgets
-#' @param server.default default character value for "SQL Server Name";
-#'   default is \code{NULL}. If \code{NULL}, default in module is
-#'   \code{paste0(\link{Sys.info}()[["nodename"]], "\\SQLEXPRESS")}
-#' @param database.default default character value for "SQL Server Database" widget;
-#'   default is "***REMOVED***"
-#' @param port.check.default default boolean value for 'Specify port number' widget;
-#'   default is \code{FALSE}
-#' @param port.default default numeric value for "Port number";
-#'   default is ***REMOVED***
-#' @param conn.default default character value for "Connection type";
-#'   default is "trusted
-#' @param uid.default default character value for "User"; default is "sa"
-#' @param pwd.default default character value for "Password";
-#'   default is \code{NULL}
+#' @param server.default default character value for "Server Hostname" widget;
+#'   default is `NULL`. If `NULL`, the module defaults to
+#'   `paste0(Sys.info()[["nodename"]], "\\SQLEXPRESS")`
+#' @param database.default default character value for "Database" widget;
+#'   default is `NULL`
+#' @param port.default default numeric value for "Port number" widget; default
+#'   is `NULL`
+#' @param conn.default default character value for "Connection type" widget;
+#'   default is `"trusted"`
+#' @param uid.default default character value for "User" widget; default is
+#'   `NULL`
+#' @param pwd.default default character value for "Password" widget; default is
+#'   `NULL`
+#'
 #' @export
 mod_database_ui <- function(id,
                             col.width = 5,
                             server.default = NULL,
-                            database.default = "***REMOVED***",
-                            port.check.default = FALSE,
-                            port.default = ***REMOVED***,
+                            database.default = NULL,
+                            port.default = NULL,
                             conn.default = "trusted",
-                            uid.default = "sa",
+                            uid.default = NULL,
                             pwd.default = NULL) {
   ns <- NS(id)
 
   if (is.null(server.default))
     server.default <- paste0(Sys.info()[["nodename"]], "\\SQLEXPRESS")
+
+  port.check.default <- !is.null(port.default)
+  drivers.list <- odbc::odbcListDrivers() %>%
+    filter(attribute == "SQLLevel", value > 0) %>%
+    pull(name) %>%
+    unique()
+
+  if (!(amlrDatabases::amlr.driver %in% drivers.list))
+    warning("The preferred driver ('", amlrDatabases::amlr.driver,
+            "') is not available on this machine")
 
   # assemble UI elements
   tagList(
@@ -47,16 +56,27 @@ mod_database_ui <- function(id,
         condition = "input.db_conn == 'other'", ns = ns,
         box(
           width = 12,
-          textInput(ns("db_other_server"), tags$h5("SQL Server Name"),
-                    value = server.default),
-          textInput(ns("db_other_database"), tags$h5("SQL Server Database"),
-                    value = database.default),
-          checkboxInput(ns("db_other_port_check"), "Specify port number",
-                        value = port.check.default),
-          conditionalPanel(
-            condition = "input.db_other_port_check == true", ns = ns,
-            numericInput(ns("db_other_port"), tags$h5("Port number"),
-                         value = port.default)
+          fluidRow(
+            column(6, textInput(ns("db_other_server"), tags$h5("Server hostname"),
+                                value = server.default)),
+            column(6, textInput(ns("db_other_database"), tags$h5("Database"),
+                                value = database.default)),
+            column(6, selectInput(ns("db_other_driver"), tags$h5("Driver"),
+                                  choices = drivers.list,
+                                  selected = amlrDatabases::amlr.driver))
+          ),
+          fluidRow(
+            column(6, checkboxInput(ns("db_other_port_check"),
+                                    "Specify port number",
+                                    value = port.check.default)),
+            column(
+              width = 6,
+              conditionalPanel(
+                condition = "input.db_other_port_check == true", ns = ns,
+                numericInput(ns("db_other_port"), tags$h5("Port number"),
+                             value = port.default)
+              )
+            )
           ),
           radioButtons(ns("db_other_conn"), tags$h5("Connection type"),
                        choices = c("Trusted connection" = "trusted",
@@ -64,8 +84,12 @@ mod_database_ui <- function(id,
                        selected = conn.default),
           conditionalPanel(
             condition = "input.db_other_conn == 'login'", ns = ns,
-            textInput(ns("db_other_uid"), tags$h5("User"), value = uid.default),
-            textInput(ns("db_other_pwd"), tags$h5("Password"), value = pwd.default)
+            fluidRow(
+              column(6, textInput(ns("db_other_uid"), tags$h5("User"),
+                                  value = uid.default)),
+              column(6, textInput(ns("db_other_pwd"), tags$h5("Password"),
+                                  value = pwd.default))
+            )
           ),
           actionButton(ns("db_other_action"), "Connect to other database")
         )
@@ -76,31 +100,26 @@ mod_database_ui <- function(id,
 
 #' @name mod_database
 #'
-#' @param pool.list a (named) list of pool objects created by \code{\link[pool]{dbPool}}.
+#' @param pool.list a (named) list of pool objects created by [pool::dbPool()].
 #'   The names of the objects in this list will appear as
-#'   'database connection' \code{\link[shiny]{radioButtons}} options.
-#' @param db.driver character; name of driver used to connect to the databases
-#' @param db.selected character; the initial selected value for the
-#'   database connection widget (\code{output$db_conn_uiOut}).
-#'   Passed to \code{selected} argument of \code{\link[shiny]{radioButtons}}.
-#'   Must be \code{NULL}, or the name of a connection in \code{pool.list}
+#'   'database connection' [shiny::radioButtons()] options.
+#' @param filedsn character; default is `NULL.` The file path to a DSN file with a
+#'   database connection. If not `NULL`, Tamatoa will try to try to establish a
+#'   database connection using [pool::dbPool]. See 'Details' for an example
 #'
 #' @details
-#' This module allows users to connect to the database of their choice:
-#' any already-created connections
-#' (e.g., to the ***REMOVED*** production database on the SWC server)
-#' or a database specified by the user.
+#' This module allows users to connect to the database of their choice: any
+#' already-created connections (e.g., a connection specified by a filedsn in a
+#' parent function) or a database specified by the user.
 #'
-#' If the user specifies their own database,
-#' then this module creates the pool object.
+#' If the user specifies their own database, then this module creates that pool
+#' object.
 #'
 #' @returns
 #' Returns a reactive of the pool connection specified by the user
 #'
 #' @export
-mod_database_server <- function(id, pool.list = list(),
-                                db.driver = "ODBC Driver 18 for SQL Server",
-                                db.selected = NULL) {
+mod_database_server <- function(id, pool.list = list(), filedsn = NULL) {
   moduleServer(
     id,
     function(input, output, session) {
@@ -109,38 +128,55 @@ mod_database_server <- function(id, pool.list = list(),
         other = FALSE
       )
 
-      encrypt.driver <- "ODBC Driver 18 for SQL Server"
-      trusted.connection <- if_else(db.driver == encrypt.driver, "Yes", "TRUE")
-      trusted.connection.no <- if_else(db.driver == encrypt.driver, "No", "FALSE")
+      # Connect to database specified by filedsn argument, if specified
+      if (!is.null(filedsn)) {
+        pool.filedsn <- try(pool::dbPool(odbc::odbc(), filedsn = filedsn),
+                            silent = TRUE)
+
+        if (isTruthy(pool.filedsn)) {
+          onStop(function() {
+            if (isTruthy(pool.filedsn))
+              if (dbIsValid(pool.filedsn)) {
+                print("closing2")
+                poolClose(pool.filedsn)
+              }
+          })
+
+          pool.list <- c(pool.list, `filedsn argument` = pool.filedsn)
+
+        } else {
+          warning("Unable to make a database connection using ",
+                  "the provided filedsn. Connection will be ignored",
+                  immediate. = TRUE)
+        }
+      }
 
 
       #----------------------------------------------------------------------------
-      # Connect to database
+      ### Render the widget to select the Database connection
+      output$db_conn_uiOut <- renderUI({
+        choices.list <- c(names(pool.list), "other")
+        names(choices.list) <- c(names(pool.list), "Other")
+
+        radioButtons(
+          session$ns("db_conn"), "Select database connection",
+          choices = choices.list
+        )
+      })
+
+      #----------------------------------------------------------------------------
+      #
 
       ### Close other database
       db_other_close <- function() {
         if (isTruthy(vals.db$pool) && vals.db$other) {
           if (dbIsValid(vals.db$pool)) {
+            print("closing1")
             poolClose(vals.db$pool)
             vals.db$pool <- NULL
           }
         }
       }
-
-      ### Database connection widget based on input
-      output$db_conn_uiOut <- renderUI({
-        validate(
-          need(is.null(db.selected) || (db.selected %in% names(pool.list)),
-                "db.selected is not a named item in pool.list")
-        )
-        choices.list <- c(names(pool.list), "other")
-        names(choices.list) <- c(names(pool.list), "Other")
-
-        radioButtons(
-          session$ns("db_conn"), tags$h5("Select database connection"),
-          choices = choices.list, selected = db.selected
-        )
-      })
 
       ### Default databases
       observeEvent(input$db_conn, {
@@ -151,9 +187,15 @@ mod_database_server <- function(id, pool.list = list(),
         vals.db$pool <- pool.list[[req(input$db_conn)]]
       })
 
-      ### Other database, on button click
+      ### Connect to an 'other' database, on button click
       observeEvent(input$db_other_action, {
         db_other_close()
+
+        db.driver <- input$db_other_driver
+        encrypt.driver <- "ODBC Driver 18 for SQL Server"
+        trusted.connection <- if_else(db.driver == encrypt.driver, "Yes", "TRUE")
+        trusted.connection.no <- if_else(db.driver == encrypt.driver, "No", "FALSE")
+
 
         db.other.conn <- if (input$db_other_conn == "trusted") {
           list(Trusted_Connection = trusted.connection)
@@ -192,15 +234,13 @@ mod_database_server <- function(id, pool.list = list(),
                      "specified the correct connection arguments?"))
         )
 
-        db.query <- dbGetQuery(
-          vals.db$pool, "SELECT @@servername, DB_NAME(), SYSTEM_USER"
-        )
+        con <- pool::localCheckout(vals.db$pool)
+        info <- pool::dbGetInfo(con)
 
         data.frame(
-          Label = c("Driver", "Server", "Database", "User"),
-          Value = unlist(
-            c(db.driver, db.query[[1]], db.query[[2]], db.query[[3]])
-          )
+          Label = c("Server", "Database", "User", "Driver", "Driver Version"),
+          Value = unlist(info[c("servername", "dbname", "username",
+                                "drivername", "driver.version")])
         )
       })