+++ R/round_logical.R
@@ -2,14 +2,14 @@
 #'
 #' Always round fives up
 #'
-#' @param x numeric; see \code{\link[base]{round}}
-#' @param digits numeric; see \code{\link[base]{round}}
+#' @param x numeric; see [base::round()]
+#' @param digits numeric; see [base::round()]
 #'
-#' @details R's default rounding function uses the IEC 60559 standard ('go to
-#' the even digit'; see \code{\link[base]{round}} for more details). While
-#' this is useful for all of the reasons outlined in the standard, sometimes
-#' AMLR scientists want to reproduce 'logical' or 'traditional' rounding,
-#' meaning rounding up from 5.
+#' @details
+#' R's default rounding function uses the IEC 60559 standard ('go to the even
+#' digit'; see [base::round()] for more details). While this is useful for all
+#' of the reasons outlined in the standard, sometimes AMLR scientists want to
+#' reproduce 'logical' or 'traditional' rounding, meaning rounding up from 5.
 #'
 #' This function is implemented from \url{https://stackoverflow.com/a/12688836},
 #' and is the same as the \code{janitor} function