+++ data-raw/data_vars.R
@@ -1,0 +1,3 @@
+## Preferred database driver
+amlr.driver <- "ODBC Driver 18 for SQL Server"
+usethis::use_data(amlr.driver, overwrite = TRUE)